build_image: Only copy portage files if installed via a non chromeos-dev rdep
The "fixup" only needs to happen if sys-apps/portage was installed under
/usr/local. If sys-apps/portage gets pulled in via a non chromeos-dev
dependency, say chromeos or a board's bsp, we don't need to do anything.

BUG=chromium:311764
TEST=build_image --board=beaglebone dev test
TEST=build_image --board=lumpy dev test
TEST=build_image --board=daisy dev test
TEST=verify make.globals exists under /usr/share/portage/config on the rootfs.

Reviewed-on: https://chromium-review.googlesource.com/191139
Tested-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: Gaurav Shah <gauravsh@chromium.org>

+++ build_library/dev_image_util.sh
@@ -46,9 +46,17 @@ install_dev_packages() {
   sudo ln -s /usr/local/usr/lib/debug "${root_fs_dir}/usr/lib/debug"
 
   # Install the bare necessary files so that the "emerge" command works
-  sudo cp -a ${root_dev_dir}/share/portage ${root_fs_dir}/usr/share
+  local portage_make_globals_path="/usr/share/portage/config/make.globals"
+  if [[ ! -f "${root_fs_dir}/${portage_make_globals_path}" ]]; then
+    # We only need to do this if portage was installed as part of the dev
+    # install.
+    # Note: The check needs to be for a file path installed by sys-apps/portage,
+    # just checking for existence of /usr/share/portage isn't sufficient as
+    # other packages may install files in there.
+    sudo cp -a ${root_dev_dir}/share/portage ${root_fs_dir}/usr/share
+  fi
   sudo sed -i s,/usr/bin/wget,wget, \
-    ${root_fs_dir}/usr/share/portage/config/make.globals
+    ${root_fs_dir}/${portage_make_globals_path}
 
   sudo mkdir -p ${root_fs_dir}/etc/make.profile
